requirements.txt
anyio==3.4.0
asgiref==3.4.1
asyncio==3.4.3
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
dnspython==2.1.0
fastapi==0.70.0
google-api-core==2.2.2
google-api-python-client==2.32.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
h11==0.12.0
httplib2==0.20.2
idna==3.3
motor==2.5.1
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pymongo==3.12.2
pyparsing==3.0.6
python-dotenv==0.19.2
requests==2.26.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
uvicorn==0.15.0
